Build on macOS, and fix four compiler warnings
Three things stopped the tree building under clang on an Apple Silicon
Mac, and nothing in it had ever had Darwin handling.

libevent: Homebrew installs outside the compiler's default search path
(/opt/homebrew on Apple Silicon, /usr/local on Intel), so event2/event.h
was not found. setup.py now resolves it -- LIBEVENT_PREFIX, then
pkg-config, then `brew --prefix libevent`, then the standard Homebrew
prefixes probed for the header. It never raises: with nothing found it
returns empty lists and lets the compiler emit its own clearer error.
Linux is unaffected, since the distro package is already on the search
path and pkg-config reports no -I is needed. The _filament.io extension
also gained library_dirs; it links event_pthreads/event_core but passed
no libevent paths at all, which only worked because Linux did not need
any.

Fiber-switch assembly: the top-level __asm__ blocks used ELF-only
directives that Apple's assembler rejects outright (.hidden, .type,
.size). Mach-O also prefixes C symbols with an underscore, so the
symbol would not have linked even once those were accepted -- worth
noting because that failure would have come later, at link time, and
looked unrelated. Both arch blocks now select the spelling on __APPLE__
via macros; the instruction bodies are untouched. Verified the ELF path
is byte-identical and nm still shows the symbol local and absent from
the dynamic table, i.e. .private_extern and .hidden agree.

Also four warnings, one of which was hiding a real inconsistency:

_NUM_RESOLVER_METHODS expanded to a bare sizeof expression (size_t), so
every comparison against a signed loop counter or the 'long' index
warned under -Wsign-compare -- three sites. Fixed at the definition
rather than the comparisons, because one of those sites needs its index
signed for an `idx < 0` check to mean anything. It stays an integer
constant expression, so it is still valid as an array dimension.

FIL_SCHED_EVENT_FREELIST_MAX was defined twice with DIFFERENT values --
256 in the header, 2048 in fil_scheduler.c. The .c value won there (it
comes after the include), so behaviour was correct, but any other
translation unit including the header would have silently used 256. The
two comments even justify different numbers, so the cap was raised and
the header missed. Consolidated on the documented 2048.

The other two are in the vendored greenlet, and both trace back to the
private-stack fiber core rather than to anything upstream. clang reports
them where gcc does not only because the normal build does not pass
-Wall; adding it reproduces both on Linux/gcc too.

An unused variable in green_new: the UserGreenlet the constructor
returns is used solely by the assert that it matches o->pimpl, and NDEBUG
compiles that away. Kept the assert, added a (void) cast.

An unused function, slp_switch: with VGL_FIBER on, g_switchstack swaps
stack pointers with fil_fiber_asm_switch and the classic switch is never
called. The platform header still has to be included -- it is also what
defines STACK_MAGIC and the register-save macros -- so the definition
stays and is marked unused, via a new GREENLET_MAYBE_UNUSED next to
GREENLET_NOINLINE. Unlike the existing UNUSED(x) it preserves the name,
so the classic core's real call site still compiles; that branch keeps
the old spelling verbatim and was rebuilt under -Wall to confirm it.
greenlet_slp_switch.hpp now includes greenlet_cpython_compat.hpp
explicitly, since that is what defines VGL_FIBER.

src/ and vendor/greenlet/ now build with no warnings under -Wall
-Wsign-compare, under either switching core.

+++ src/socket/fil_socket.c
@@ -324,7 +324,12 @@ static char *_RESOLVER_METHOD_NAMES[] = {
     "getnameinfo",
     NULL,
 };
-#define _NUM_RESOLVER_METHODS ((sizeof(_RESOLVER_METHOD_NAMES) / sizeof(char *)) - 1)
+/* Cast to int: this is compared against signed loop counters and against a
+ * 'long' index below, and a bare sizeof expression is size_t -- which makes
+ * every one of those comparisons signed/unsigned and warns under
+ * -Wsign-compare.  It stays an integer constant expression, so it is still
+ * valid as an array dimension. */
+#define _NUM_RESOLVER_METHODS ((int)((sizeof(_RESOLVER_METHOD_NAMES) / sizeof(char *)) - 1))
 static PyObject *_RESOLVER_METHODS[_NUM_RESOLVER_METHODS];
 
 static inline int _copy_resolver_methods(PyObject *resolver)
